+++ images/ananke-base/assets/install/rtc.sh
@@ -3,7 +3,7 @@
 source /opt/conda/etc/profile.d/conda.sh
 
 conda activate jhub
-pip install jupyter-collaboration==2.0.0a4
+pip install jupyter-collaboration==2.0.2
 
 # disabel RTC for all users
 jupyter labextension disable @jupyter/collaboration-extension